[#51] 개별 룩북 코디 조회 api 추가
#51

+++ Codive/DIContainer/LookBookDIContainer.swift
@@ -73,7 +73,7 @@ final class LookBookDIContainer {
     
     // MARK: - Specific LookBook
     func makeSpecificLookBookViewModel(
-        lookbookId: Int,
+        lookbookId: Int64,
         lookbookTitle: String = ""
     ) -> SpecificLookBookViewModel {
         return SpecificLookBookViewModel(
@@ -85,7 +85,7 @@ final class LookBookDIContainer {
     }
     
     func makeSpecificLookBookView(
-        lookbookId: Int,
+        lookbookId: Int64,
         lookbookTitle: String = ""
     ) -> SpecificLookBookView {
         return SpecificLookBookView(
